Add Vue Test Utils (#72)
The Vue Test Utils are "a set of utility functions aimed to simplify
testing Vue.js components". This package allows us to unit test 
the Vue components of Numberscope.

Also, adds complete examples of a Vue component test (SeqGetter.vue)
and of a Vue view (Home.vue). The tests may be run with `npm run test:vue`.
This works toward, but does not fully complete, a resolution of issue #25.