chore: Rename Title
We rename the title from "newnumberscope" to "Numberscope". The old
title was obtained from the "name" field in `package.json`. I'm not a
fan of this method since the "name" field in `package.json` is supposed
to be lowercase. Thus, we change the "name" in `package.json` to be
"frontscope" to match the repo's name.

Related to https://github.com/numberscope/frontscope/pull/88 in that
it's a minor cosmetic thing that will make the frontend look a bit
nicer.