fix: [WIP] Add Work on Integrating Everything
In this commit, we can `npm run dev` and have a somewhat broken front
end. This is an improvement from the previous commit where we couldn't
even see the front end because of errors. We still need to fix the
front end so that it behaves the same as the Vue 2 version of
frontscope.