docs: Remove Info on Configuring Vite
One of the "killer features" of Vite is that it's supposed to be
zero-config. I don't know why (at this point) we'd want to configure
Vite, and I agree with Glen's comment
https://github.com/numberscope/frontscope/pull/123#issuecomment-1165783514
that configuring Vite isn't important enough to be included in the
README. It's a vestigial piece of documentation from initializing the
Vue app.