docs: Remove Info on Configuring Vite
The link to the Vite docs was auto-added when I initialized the Vue
project. I don't see why (at this point) we'd want to configure Vite.
It's supposed to be "zero-config". Also, even if we needed to keep this
information around, I don't think the README would be the right place to
house it.