doc: Harmonize guidance on git, cloning, and other workflows. (#463)
* doc: Harmonize guidance on git, cloning, and other workflows.

Resolves #221.

THere is still a signifcant amount of repetition/overlap among
CONTRIBUTING.md, and in the doc/ directory, onboarding.md,
running-from-source.md, and working-with-git-and-github.md. I wasn't
sure how we might compartmentalize these, because they are at
different levels of detail and for different audiences.

Since they are all now consistent with each other, I am not too fussed
by the repetitiveness. Maybe it's OK to give potential contributors multiple
places to find the information they need. But if the reviewer has suggestions
about how to reduce the redundancy, I am also open to making more changes.

* doc: Harmonize gitting-it-right with all of the other git/workflow docs

* doc: Changes pursuant to PR review