Numberglyph overhaul (#571)
Fixes a number of small bugs with NumberGlyph (an off-by-one error, a race condition in multiple instances of constructing the palette from factorizations), and adjusts the layout to fill the full width of the available rectangle of screen space. Adds a Featured Specimen using NumberGlyph. Also (unrelatedly) restores the ability for documentation pages to use LaTeX markup for formulas.

Resolves #282.
Resolves #365.
Resolves #396.
Resolves #413.
Resolves #451.
---------

Co-authored-by: Glen Whitney <glen@studioinfinity.org>